+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2023-07-16
+### Added
+- PinchInputForwarder Component to forward drag event to the PinchInputDetector from any raycast blocking components.
+- OnPinchStart/OnPinchEnd Unity events in PinchInputDetector Component.
+
 ## [1.0.6] - 2023-04-19
 ### Fixed
 - Fix an bug where the content zoom out become out of boundary when movement type is Clamped.
@@ -37,6 +42,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-02-12
 ### Added
-- PinchableScrollRect Component
-- IPinchHandler Interface for extending pinching behaviours
-- PinchInputDetector Component to convert pointer event to pinch event
+- PinchableScrollRect Component.
+- IPinchHandler Interface for extending pinching behaviours.
+- PinchInputDetector Component to convert pointer event to pinch event.